refactor: `/write` 글쓰기 페이지 컴포넌트 디렉터리 생성 및 글쓰기 페이지 내부 `Header` 컴포넌트 분리
